Moved code to src dir & brought setup files to python3 standards
