+++ setup.cfg
@@ -1,6 +1,118 @@
+##########################
+# Setup.py Configuration #
+##########################
+[metadata]
+name = opendrop
+version = 0.11.0
+description = An open Apple AirDrop implementation
+long_description = file: README.md
+long_description_content_type = text/markdown
+
+# URLs associated with the project
+url = https://owlink.org
+download_url = FIXME
+project_urls =
+    Bug Tracker = FIXME
+    Source Code = https://github.com/seemoo-lab/opendrop
+    Documentation = FIXME
+    Research Paper = https://usenix.org/conference/usenixsecurity19/presentation/stute
+
+# Author information
+author = The Open Wireless Link Project
+author_email = mstute@seemoo.tu-darmstadt.de
+maintainer = The Open Wireless Link Project
+maintainer_email = mstute@seemoo.tu-darmstadt.de
+
+# License Information
+license = See LICENSE file
+license_file = LICENSE
+
+# Search tags
+classifiers =
+    Environment :: Console
+    License :: OSI Approved :: GNU General Public License v3 (GPLv3)
+    Operating System :: MacOS
+    Operating System :: POSIX :: Linux
+    Natural Language :: English
+    Programming Language :: Python
+    Programming Language :: Python :: 3.7
+    Programming Language :: Python :: 3.6
+    Programming Language :: Python :: 3 :: Only
+
+keywords = cli
+
+[options]
+install_requires =
+    Pillow
+    ctypescrypto
+    fleep
+    ifaddr
+    libarchive-c
+    requests
+    requests_toolbelt
+    click
+    zeroconf>=0.24.2
+
+# Default options
+zip_safe = false
+python_requires = >=3.6
+
+# Package Data
+include_package_data = True
+package_data =
+    opendrop = opendrop/certs/*.pem
+
+# Where to find my code
+packages = find:
+package_dir =
+    = src
+
+[options.packages.find]
+where = src
+
+[options.extras_require]
+docs =
+    sphinx
+    sphinx-rtd-theme
+    sphinx-click
+    sphinx-autodoc-typehints
+
+[options.entry_points]
+console_scripts =
+    opendrop = opendrop.cli:main
+
+######################
+# Doc8 Configuration #
+# (doc8.ini)         #
+######################
+[doc8]
+max-line-length = 120
+
+##########################
+# Coverage Configuration #
+# (.coveragerc)          #
+##########################
+[coverage:run]
+branch = True
+source = opendrop
+
+[coverage:report]
+show_missing = True
+exclude_lines =
+    def __str__
+    def __repr__
+
+########################
+# Flake8 Configuration #
+# (.flake8)            #
+########################
 [flake8]
 max-line-length = 127
 
+######################
+# YAPF Configuration #
+# (.style.yapf)      #
+######################
 [yapf]
 based_on_style = pep8
 column_limit = 127
