+++ tests/test_client.py
@@ -1,5 +1,5 @@
-from opendrop.client import AirDropBrowser
-from opendrop.config import AirDropConfig
+from src.opendrop.client import AirDropBrowser
+from src.opendrop.config import AirDropConfig
 
 
 def get_loopback():
